build.csh now has an argument
If arg1 is Dockerfile, then build for ARW (use 8th try)
If arg1 is Dockerfile-NMM, then build for HWRF (use 6th try)

ARW requires GNU 9 to build the RRTMG-derivative physics schemes with
"configure -d". Unfortunately, HWRF is not able to run successfully with
GNU 9.

Therefore we have two containers: one for ARW, and the other for HWRF.