Update for jenkins, stop removing image (1/2)
When there are several docker commands running on the same machine,
we cannot end each of the tests (such as test_001s.csh) with the
clean up duties that include removing the container. This is a shared
resource between serial, openmp, and mpi.

As there used to only be a single.csh (a start-up that built the image),
we now have two "single" scripts.
1. single_init.csh is IDENTICAL to the original single.csh
2. single_end.csh is a new script. It contains the several lines that
are in each of the test_0xx.csh files that remove the image (and
prune, and do that magical system df).

An example ordering of the scripts could be:

single_init.csh
test_001.csh &
test_002.csh &
test_003.csh &
wait
single_end.csh

The only modified file is:
build.csh