Conventionalize single_init.csh and single_end.csh args
The scripts now use the necessary input data directly, instead of
using the "oh this is from a Dockerfile, so lets assume that the
image name is wrf_regtest", we now bring in two args for
single_init.csh: the name of the Dockerfile, and the name of the
iamge to generate.

This is consistent with single_end.csh, where we now pass in the
name of the image to remove.

Changes to be committed:
modified:   build.csh