Move chemistry tet to be the first test
The purpose for this re-ordering is due to time constraints. The chemistry
tests take more than 3x as long to build as the other ARW tests. For $
concerns, all tests were moved to a cheaper (smaller) AWS instance. This
smaller machine type had fewer proessors, and therefore compiled the chem
tests too slowly. The decision was to bump up the machine type, but only
for chemistry job. The easiest way to handle the split machine assignment in
the groovy scripting is to make the chemistry tests the first job.

modified:   build.csh