Make the feature test the second in the list
To test this on the MMM Classroom machines, I need to have the feature test
one of the early tests, since the scripts are broken apart. The actual order
does not matter, as long as the first test is chemistry (which requires a
larger machine on AWS).

modified:   build.csh